refactor: ホーム画面の検索・初期データ取得ロジックをカスタムフックに分離
- useEffectをカスタムフックに分離
